oim/nvidia/nemotron-nano-12b-v2-vl/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "oim-nemotron-nano-12b-v2-vl"
version = "0.1.0"
description = "Open replacement for nemotron-nano-12b-v2-vl packaged for uv."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Open NV Ingest Maintainers" }]
dependencies = [
  "fastapi==0.124.4",
  "numpy==2.3.5",
  "opentelemetry-api==1.39.1",
  "opentelemetry-exporter-otlp-proto-http==1.39.1",
  "opentelemetry-sdk==1.39.1",
  "oim-common>=0.1.0",
  "pillow==12.0.0",
  "prometheus-client==0.23.1",
  "pydantic-settings==2.12.0",
  "pydantic==2.12.5",
  "tenacity==9.0.0",
  "torch==2.8.0",
  "transformers==4.57.3",
  "tritonclient[http]==2.63.0; platform_system == \"Linux\"",
  "tritonclient[http]==2.60.0; platform_system != \"Linux\"",
  "uvicorn[standard]==0.38.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "ruff>=0.7.0", "mypy>=1.13.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/nemotron_nano_12b_v2_vl"]

[tool.uv]
managed = true

[tool.uv.sources]
common = { workspace = true }
